requirements.txt
pandas 
numpy
matplotlib
seaborn
neo4j
fastapi
python-dotenv